[Feature][Server] Add data quality report feature

+++ datavines-server/src/main/java/io/datavines/server/repository/entity/ActualValues.java
@@ -24,6 +24,7 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Data;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -48,7 +49,7 @@ public class ActualValues implements Serializable {
     private String actualName;
 
     @TableField(value = "actual_value")
-    private Double actualValue;
+    private BigDecimal actualValue;
 
     @JsonFormat(pattern="yyyy-MM-dd HH:mm:ss",timezone = "GMT+8")
     @TableField(value = "data_time")
